refactor server api graph ops
